+++ src/aioprometheus/__init__.py
@@ -3,6 +3,9 @@ from .decorators import count_exceptions, inprogress, timer
 from .negotiator import negotiate
 from .renderer import render
 
+# To avoid cicular imports issues asgi must be imported after other modules
+from .asgi.middleware import MetricsMiddleware  # isort:skip
+
 # The 'pusher' and 'service'  modules must be explicitly imported by package
 # users as they depend on optional extras.